Move more exports into separate files
