+++ src/timing.js
@@ -1,0 +1,57 @@
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @flow
+ */
+'use strict';
+
+var Animated = require('./Animated');
+var AnimatedValue = require('./AnimatedValue');
+var AnimatedValueXY = require('./AnimatedValueXY');
+var AnimatedTracking = require('./AnimatedTracking');
+var TimingAnimation = require('./TimingAnimation');
+var maybeVectorAnim = require('./maybeVectorAnim');
+
+import type { CompositeAnimation, EndCallback } from './Animation';
+
+type TimingAnimationConfig =  AnimationConfig & {
+  toValue: number | AnimatedValue | {x: number, y: number} | AnimatedValueXY;
+  easing?: (value: number) => number;
+  duration?: number;
+  delay?: number;
+};
+
+var timing = function(
+  value: AnimatedValue | AnimatedValueXY,
+  config: TimingAnimationConfig,
+): CompositeAnimation {
+  return maybeVectorAnim(value, config, timing) || {
+    start: function(callback?: ?EndCallback): void {
+      var singleValue: any = value;
+      var singleConfig: any = config;
+      singleValue.stopTracking();
+      if (config.toValue instanceof Animated) {
+        singleValue.track(new AnimatedTracking(
+          singleValue,
+          config.toValue,
+          TimingAnimation,
+          singleConfig,
+          callback
+        ));
+      } else {
+        singleValue.animate(new TimingAnimation(singleConfig), callback);
+      }
+    },
+
+    stop: function(): void {
+      value.stopAnimation();
+    },
+  };
+};
+
+module.exports = timing;
